Add toString method to State, add word length constraint
The Polish dictionary for scrabble contains nearly 3.5mln words, some of which are words with 10+ letters which will be very rarely used in the real game scenario. As it's not possible to load all of them into memory, because that causes the heap to run out of space, the word length constraint is introduced.